--- sdk/src/main/java/com/stytch/sdk/StytchEnvironment.kt
@@ -1,6 +1,0 @@
-package com.stytch.sdk
-
-public enum class StytchEnvironment {
-    LIVE,
-    TEST
-}
