--- sdk/src/main/java/com/stytch/sdk/StytchError.kt
@@ -1,9 +1,0 @@
-package com.stytch.sdk
-
-public enum class StytchError(public val messageId: Int) {
-    InvalidEmail(R.string.stytch_error_invalid_input),
-    Connection(R.string.stytch_error_no_internet),
-    Unknown(R.string.stytch_error_unknown),
-    InvalidMagicToken(R.string.stytch_error_invalid_magic_token),
-    InvalidConfiguration(R.string.stytch_error_bad_token)
-}
