--- sdk/src/main/java/com/stytch/sdk/StytchLoginMethod.kt
@@ -1,6 +1,0 @@
-package com.stytch.sdk
-
-public enum class StytchLoginMethod {
-    LoginOrSignUp,
-    LoginOrInvite
-}
