--- sdk/src/main/java/com/stytch/sdk/api/StytchResult.kt
@@ -1,10 +1,0 @@
-package com.stytch.sdk.api
-
-public data class StytchResult(
-    public val userId: String,
-    public val requestId: String,
-){
-    override fun toString(): String {
-        return "StytchResult(userId='$userId', requestId='$requestId')"
-    }
-}
