--- sdk/src/main/java/com/stytch/sdk/api/requests/CreateUserRequest.kt
@@ -1,8 +1,0 @@
-package com.stytch.sdk.api.requests
-
-import com.google.gson.annotations.SerializedName
-
-public class CreateUserRequest(
-    @SerializedName("email") public val email: String,
-): BasicRequest() {
-}
