--- sdk/src/main/java/com/stytch/sdk/api/requests/SendEmailVerificationRequest.kt
@@ -1,4 +1,0 @@
-package com.stytch.sdk.api.requests
-
-public class SendEmailVerificationRequest(): BasicRequest() {
-}
