--- sdk/src/main/java/com/stytch/sdk/api/responses/BasicResponse.kt
@@ -1,4 +1,0 @@
-package com.stytch.sdk.api.responses
-
-public open class BasicResponse {
-}
