--- sdk/src/main/java/com/stytch/sdk/api/responses/DeleteUserResponse.kt
@@ -1,4 +1,0 @@
-package com.stytch.sdk.api.responses
-
-public class DeleteUserResponse: BasicResponse() {
-}
