--- sdk/src/main/java/com/stytch/sdk/api/responses/SendEmailVerificationResponse.kt
@@ -1,8 +1,0 @@
-package com.stytch.sdk.api.responses
-
-import com.google.gson.annotations.SerializedName
-
-public class SendEmailVerificationResponse(
-    @SerializedName("user_id") public val user_id: String,
-) : BasicResponse() {
-}
