+++ sdk/src/main/java/com/stytch/sdk/core.kt
@@ -1,9 +1,6 @@
 package com.stytch.sdk
 
 import android.net.Uri
-import com.stytch.sdk.api.StytchResult
-import com.stytch.sdk.helpers.Constants
-import com.stytch.sdk.helpers.StytchFlowManager
 
 public class Stytch private constructor() {
     internal var config: StytchConfig? = null
@@ -83,4 +80,100 @@ public class Stytch private constructor() {
         public fun onMagicLinkSent(email: String)
     }
 
+}
+
+internal class StytchConfig private constructor(
+    val projectID: String,
+    val secret: String,
+    val deepLinkScheme: String,
+    val deepLinkHost: String,
+    val verifyEmail: Boolean,
+    var uiCustomization: StytchUICustomization
+) {
+
+    public class Builder {
+
+        private var deepLinkScheme: String = "app"
+        private var deepLinkHost: String = "stytch.com"
+        private var uiCustomization: StytchUICustomization = StytchUICustomization()
+        private var verifyEmail: Boolean = false
+        private var projectID: String = ""
+        private var secret: String = ""
+
+        fun withDeepLinkScheme(scheme: String): Builder {
+            deepLinkScheme = scheme
+            return this
+        }
+
+        fun withDeepLinkHost(host: String): Builder {
+            deepLinkHost = host
+            return this
+        }
+
+        fun withCustomization(UICustomization: StytchUICustomization): Builder {
+            this.uiCustomization = UICustomization
+            return this
+        }
+
+        fun withVerifyEmail(verifyEmail: Boolean): Builder {
+            this.verifyEmail = verifyEmail
+            return this
+        }
+
+        fun withAuth(projectID: String, secret: String): Builder {
+            this.projectID = projectID
+            this.secret = secret
+            return this
+        }
+
+        fun build(): StytchConfig {
+            return StytchConfig(
+                projectID,
+                secret,
+                deepLinkScheme,
+                deepLinkHost,
+                verifyEmail,
+                uiCustomization,
+            )
+        }
+    }
+
+}
+
+public enum class StytchEnvironment {
+    LIVE,
+    TEST
+}
+
+public enum class StytchError(public val messageId: Int) {
+    InvalidEmail(R.string.stytch_error_invalid_input),
+    Connection(R.string.stytch_error_no_internet),
+    Unknown(R.string.stytch_error_unknown),
+    InvalidMagicToken(R.string.stytch_error_invalid_magic_token),
+    InvalidConfiguration(R.string.stytch_error_bad_token)
+}
+
+
+public class StytchEvent private constructor(
+    public val type: String,
+    public val created: Boolean,
+    public val userId: String,
+) {
+
+    public companion object {
+        private const val USER_EVENT = "user_event"
+
+        public fun userCreatedEvent(userId: String): StytchEvent {
+            return StytchEvent(USER_EVENT, true, userId)
+        }
+
+        public fun userFoundEvent(userId: String): StytchEvent {
+            return StytchEvent(USER_EVENT, false, userId)
+        }
+    }
+}
+
+public enum class StytchLoginMethod {
+    LoginOrSignUp,
+    LoginOrInvite
 }
