--- sdk/src/main/java/com/stytch/sdk/exceptions/BlankWarningException.kt
@@ -1,4 +1,0 @@
-package com.stytch.sdk.exceptions
-
-internal class BlankWarningException(val errorMessage: String?) : Exception() {
-}
