--- sdk/src/main/java/com/stytch/sdk/exceptions/EmailNotFoundException.kt
@@ -1,4 +1,0 @@
-package com.stytch.sdk.exceptions
-
-internal class EmailNotFoundException: Exception() {
-}
