--- sdk/src/main/java/com/stytch/sdk/exceptions/WarningException.kt
@@ -1,4 +1,0 @@
-package com.stytch.sdk.exceptions
-
-internal class WarningException(val errorMessageId:Int): Exception() {
-}
