--- sdk/src/main/java/com/stytch/sdk/exceptions/WrongMagicLinkException.kt
@@ -1,4 +1,0 @@
-package com.stytch.sdk.exceptions
-
-internal class WrongMagicLinkException: Exception() {
-}
