+++ sdk/src/test/java/com/stytch/sdk/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.stytch.sdk
 
+import org.junit.Assert
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
@@ -10,8 +9,15 @@ import org.junit.Assert.*
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 internal class ExampleUnitTest {
+
     @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
+    fun `assert Stytch fails with uninitialized property access exception when attempting login without setting up configuration`() {
+        try {
+            Stytch.instance.login("Hi!")
+        } catch (e: UninitializedPropertyAccessException) {
+            return
+        }
+
+        Assert.fail()
     }
 }
